LayoutTests/fast/canvas/CanvasRenderingContext2DSettings-pixelFormat-enabled-expected.txt

Test that the colorType member of CanvasRenderingContext2DSettings is supported when CanvasColorTypeEnabled=true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSettings.colorType is "unorm8"
PASS testSettings.colorType is "unorm8"
PASS testSettings.colorType is "float16"
PASS document.createElement("canvas").getContext("2d", { colorType: "foo" }) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
